ready hw2

+++ src/com/boshko/lesson2/hw2/Calculator.java
@@ -4,10 +4,12 @@ public class Calculator {
     public static double sum(double a, double b) {
         return a + b;
     }
-    public static double sum(long a, long b) {
+
+    public static long sum(long a, long b) {
         return a + b;
     }
-    public static double sum(int a, int b) {
+
+    public static int sum(int a, int b) {
         return a + b;
     }
 
@@ -15,21 +17,25 @@ public class Calculator {
     public static double subtraction(double a, double b) {
         return a - b;
     }
-    public static double subtraction(long a, long b) {
+
+    public static long subtraction(long a, long b) {
         return a - b;
     }
-    public static double subtraction(int a, int b) {
+
+    public static int subtraction(int a, int b) {
         return a - b;
     }
 
     //-----------------------------------------------------
 
     public static double division(double a, double b) {
-        return (double) a / b;
+        return a / b;
     }
+
     public static double division(long a, long b) {
         return (double) a / b;
     }
+
     public static double division(int a, int b) {
         return (double) a / b;
     }
@@ -39,10 +45,12 @@ public class Calculator {
     public static double multiplication(double a, double b) {
         return a * b;
     }
-    public static double multiplication(long a, long b) {
+
+    public static long multiplication(long a, long b) {
         return a * b;
     }
-    public static double multiplication(int a, int b) {
+
+    public static int multiplication(int a, int b) {
         return a * b;
     }
 }
